[#17] chore : constant, utill 폴더 정리
