+++ src/constant/index.js
@@ -3,3 +3,39 @@ export const CREATE_ACCOUNT = '/auth/createAccount';
 export const EMAIL_ERROR = '125';
 export const NICKNAME_ERROR = '124';
 export const ACCOUNT_ERROR = '123';
+
+
+// keyword
+export const WEEK = [
+    {id:0,day:"월요일"},
+    {id:1,day:"화요일"},
+    {id:2,day:"수요일"},
+    {id:3,day:"목요일"},
+    {id:4,day:"금요일"},
+    {id:5,day:"토요일"},
+    {id:6,day:"일요일"},
+    {id:7,day:"공휴일"},
+];
+
+export const menuItem = [
+    {   
+        id:0,
+        title:'전체'
+    },
+    {   
+        id:1,
+        title:'아우누리'
+    },
+    {   
+        id:2,
+        title:'아우미르'
+    },
+    {   
+        id:3,
+        title:'대신 전해드립니다-koreatech'
+    },
+];
+
+export const AUNURI = 'https://portal.koreatech.ac.kr';
+export const AOUMIR = 'https://dorm.koreatech.ac.kr/';
+export const KOREATECHINSTEADDELIVER = 'https://ko-kr.facebook.com/s.g.of.koreatech/';
